fix: Parse instrumented tool result messages as tool role messages (#5471)
* fix: Parse instrumented tool result messages as tool role messages

Any input message from a span that follows semantic conventions and contains a tool_call_id attribute, will have its role transformed into the "tool" ChatMessageRole.

This allows the playground to distinguish a message as one containing the result of a previously called tool.

There is no semantic convention specification otherwise to specify that a message contains the result of a tool call, and role does not have any formal meaning, so we overload its meaning in the playground

* Remove dead code